feature(tj): add MoE gradient conflict experiment reproduction README and configs
